website/backend/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
python-dotenv
passlib
